wip: save dump

+++ packages/algorithms/data-structure/bin/prebuild.mjs
@@ -1,9 +1,8 @@
-const esbuild = require("esbuild");
-const fs = require("node:fs");
-const { spawn } = require("node:child_process");
+import esbuild from "esbuild";
+import fs from "node:fs";
+import {spawn} from "node:child_process";
 
-const packageJSON = require('../package.json');
-
+import packageJSON from '../package.json' with {type: 'json'};
 
 function spawnCommand(command, args, cwd) {
     return new Promise((resolve, reject) => {
@@ -11,7 +10,7 @@ function spawnCommand(command, args, cwd) {
             cwd,
             stdio: "inherit",
             shell: false,
-            env: { ...process.env },
+            env: {...process.env},
         });
 
         child.on("close", code => {
@@ -29,9 +28,16 @@ const externals = [...Object.keys(packageJSON.dependencies || {}), ...Object.key
 
 (async function () {
 
-    fs.rmSync("lib", { recursive: true, force: true });
+    if (fs.existsSync(".temp") && fs.existsSync("lib")) 
+    {
+        console.log("skipped", packageJSON.name);
+        return;
+    }
+
+    fs.rmSync("lib", {recursive: true, force: true});
+    fs.rmSync(".temp", {recursive: true, force: true});
     await spawnCommand("tsc", ["--emitDeclarationOnly", "--declarationDir", "lib"], process.cwd());
-    fs.writeFileSync("lib/bundle.d.ts", "export * from './src';", { encoding: "utf-8" });
+    fs.writeFileSync("lib/bundle.d.ts", "export * from './src';", {encoding: "utf-8"});
 
     const esbuildInfo = await esbuild.build({
         entryPoints: ["src/index.ts"],
